requirements.txt
amqp==5.1.1
asgiref @ file:///home/conda/feedstock_root/build_artifacts/asgiref_1671589829431/work
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django @ file:///home/conda/feedstock_root/build_artifacts/django_1680693679482/work
django-cors-headers==4.0.0
django-environ==0.10.0
django-extensions==3.2.3
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
drf-spectacular==0.26.5
drf-yasg==1.21.7
face-recognition==1.3.0
face-recognition-models==0.3.0
graphviz==0.20.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.3.2
MarkupSafe==2.1.3
numpy==1.25.1
openapi-codec==1.3.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
prompt-toolkit==3.0.39
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
simplejson==3.19.1
six==1.16.0
sqlparse @ file:///home/conda/feedstock_root/build_artifacts/sqlparse_1663990453245/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
vine==5.0.0
wcwidth==0.2.8
